requirements.txt
Django
djangorestframework
psycopg2
python-dotenv
django-cors-headers
gunicorn
pytest
django-filter
drf-yasg
